+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Fixed:
 - Distinguish between reacts in chathistory vs active history
 - Channel matching on single line only
 
+Changed:
+
+- On-join topic messages disabled by default (`buffer.server_messages.topic`), instead the topic banner is enabled by default (`buffer.channel.topic_banner`)
+
 Thanks:
 
 - Contributions: @bb010g, @furudean, @englut, @luca020400